split ASCII client base out, fix expect issue
